+++ R3_VillagePeople_Mahtimokit/R3_VillagePeople_Mahtimokit.csproj
@@ -78,9 +78,7 @@
     <Compile Include="Customer_popup.Designer.cs">
       <DependentUpon>Customer_popup.cs</DependentUpon>
     </Compile>
-    <Compile Include="Db_queries.cs">
-      <SubType>Form</SubType>
-    </Compile>
+    <Compile Include="Db_queries.cs" />
     <Compile Include="Invoicing.cs">
       <SubType>Form</SubType>
     </Compile>
